requirements.txt
# TTL Monitor: TikTok Live capture via CDP
playwright>=1.40.0
Pillow>=10.0.0

# Dashboard
Flask>=2.0.0
gunicorn>=22.0.0
psutil>=5.9.0
psycopg2-binary>=2.9.0
reportlab>=4.0.0
authlib>=1.3.0
